## Profiling Analysis
before
￼<img width="1926" alt="Pasted Graphic 11" src="https://github.com/user-attachments/assets/3d333419-eea3-4ba3-a0e1-daa1670e4d52" />

after
<img width="1909" alt="Pasted Graphic" src="https://github.com/user-attachments/assets/b019b62d-5033-4ed9-9fdb-34e4e1d9e76a" />

## Methods
Below, is some analysis on the results of running the `protx diff` RPC 500 times. The diffs had a start block between MIN and MAX as defined below; and an end block no more than MAX_DIFF from the selected start block. We then perform some statistical analysis on the data.

MIN_VALUE = 1500050
MAX_VALUE = 2000050
MAX_DIFF = 50000
￼
## Statistical Analysis, outliers included

### Before
Five-Number Summary of Execution Times:
Min:    0.024492 sec
Q1:     0.124626 sec
Median: 0.243000 sec
Q3:     0.358459 sec
Max:    15.583948 sec

Mean Execution Time: 0.428296 sec
Standard Deviation: 0.933486 sec

Linear Regression Results:
y = 0.000001 * x + 0.308662
R-squared: 0.008160 (Goodness of Fit)
![Observed Data](https://github.com/user-attachments/assets/12fdd894-0f2c-4cba-9945-5e9d084c9b24)
![Pasted Graphic 6](https://github.com/user-attachments/assets/97706786-73b1-417d-9036-caaa3add1290)

### After
Five-Number Summary of Execution Times:
Min:    0.038174 sec
Q1:     0.121363 sec
Median: 0.158175 sec
Q3:     0.215866 sec
Max:    16.587903 sec

Mean Execution Time: 0.239239 sec
Standard Deviation: 0.762387 sec

Linear Regression Results:
y = 0.000001 * x + 0.151169
R-squared: 0.006918 (Goodness of Fit)
P-value: 0.063105 (Significance)
![Observed Data](https://github.com/user-attachments/assets/d50e5b5b-1a67-47c9-980c-564adab083ba)
![Observed Data](https://github.com/user-attachments/assets/cecc0394-f484-410a-ad21-1a9bbd5a1dc7)

## Statistical Analysis, outliers excluded

### Before
removed 76 data points
Five-Number Summary of Execution Times (After Outlier Removal):
Min:    0.035916 sec
Q1:     0.211060 sec
Median: 0.319278 sec
Q3:     0.357963 sec
Max:    0.572785 sec

Mean Execution Time: 0.289764 sec
Standard Deviation: 0.101140 sec

Linear Regression Results (After Outlier Removal):
y = 0.0000000199 * x + 0.286447
R-squared: 0.000496 (Goodness of Fit)
![Pasted Graphic 10](https://github.com/user-attachments/assets/1f0f6389-dc8f-48a5-80b7-9b5ccb3869dc)


### After

removed 32 data points
Five-Number Summary of Execution Times (After Outlier Removal):
Min:    0.038174 sec
Q1:     0.119880 sec
Median: 0.151724 sec
Q3:     0.205017 sec
Max:    0.355078 sec

Mean Execution Time: 0.164165 sec
Standard Deviation: 0.060919 sec

Linear Regression Results (After Outlier Removal):
y = 0.0000003119 * x + 0.111002
R-squared: 0.399298 (Goodness of Fit)

![Observed Data](https://github.com/user-attachments/assets/91e97214-6afb-4c3a-a98a-cd7e5704e724)


## How Has This Been Tested?
Ran unit tests locally; reindexing currently, going to let CI run functional tests

## Breaking Changes
Should be none; but please think through the diff specifically related to https://github.com/dashpay/dash/compare/develop...PastaPastaPasta:dash:perf-build-simplified-mn-list-diff-bls-compare-to-null?expand=1#diff-0998f8dfc4c1089e90cbaafe9607b361035b904cd103df31e3c2339a3cbf790dR480

## Checklist:
  _Go over all the following points, and put an `x` in all the boxes that apply._
- [ ] I have performed a self-review of my own code
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have added or updated relevant unit/integration/functional/e2e tests
- [ ] I have made corresponding changes to the documentation
- [x] I have assigned this pull request to a milestone _(for repository code-owners and collaborators only)_